Collapse mocking_utils into one.
